+++ app/models/RequestRow.scala
@@ -2,7 +2,7 @@ package models
 
 import play.api.libs.json._
 
-case class RequestRow(id: Long, requestId: Long, productId: Long, productorId: Long, quantity: Int, precio: Int, paid: Int, status: String)
+case class RequestRow(id: Long, requestId: Long, productId: Long, productorId: Long, quantity: Int, precio: Double, paid: Int, status: String)
 
 object RequestRow {
   implicit val RequestRowFormat = Json.format[RequestRow]
